requirements.txt
pytesseract
pdfplumber
pillow
opencv-python
Pymupdf